requirements.txt
streamlit
yfinance
pandas
plotly
google-generativeai
pypdf